[IMPROVEMENT] Cache languange when change
